+++ src/ActiveRecordHistory.php
@@ -62,7 +62,19 @@ class ActiveRecordHistory extends \yii\db\ActiveRecord
         $manager = new $this->_historyManager;
 
         return $manager->getData($attribute, $this);
-                
+    }
+
+    /**
+     * Get the records corresponding to all changes
+     * @param $attribute
+     * @return mixed
+     */
+    public function changes()
+    {
+        $manager = new $this->manager;
+        $manager->setOptions($this->managerOptions);
+
+        return $manager->getAllData($this->owner);
     }
 
 }